# finance-ui/src/services/api/models/MonitoringSummaryDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MonitoringSummaryDTO = {
    requestsPerSecond: number;
    errors4xxPerSecond: number;
    errors5xxPerSecond: number;
    p95LatencyMs: number;
    financeApiUp: boolean;
};