# finance-ui/src/services/api/models/SearchFriendRequestDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SearchFriendRequestDTO = {
    nameOrMail: string;
};