# finance-ui/src/services/api/models/UpdateSettingsReqDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateSettingsReqDTO = {
    lowBalanceEnabled?: boolean;
    lowBalanceThreshold?: number;
    newPin?: string;
};