# finance-ui/src/services/api/models/WeeklyBarDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WeeklyBarDTO = {
    day: string;
    total: number;
};